# server/common/api.yml
openapi: 3.0.1
info:
  title: crud
  description: crud
  version: 1.0.0
servers:
- url: /api/v1
...
paths:
...
  /fornitore/{id}:
    get:
      tags:
      - Fornitore
      description: Ottieni un Fornitore
      parameters:
        - name: id
          in: path
          description: id del fornitore da ottenere
          required: true
          schema:
            type: string
      responses:
        200:
          description: Restituito il Fornitore con l'id specificato
          content: {}
        404:
          description: Fornitore non trovato
          content: {}
    put:
      tags:
      - Fornitore
      description: Aggiorna un Fornitore
      parameters:
        - name: id
          in: path
          description: id del fornitore da aggiornare
          required: true
          schema:
            type: string
      requestBody:
        description: fornitore object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FornitoreBody"
        required: true
      responses:
        200:
          description: Aggiornato il fornitore con l'id specificato
          content: {}
        404:
          description: Fornitore non trovato
          content: {}
    delete:
      tags:
      - Fornitore
      description: Cancella un Fornitore
      parameters:
        - name: id
          in: path
          description: l'id del fornitore da eliminare
          required: true
          schema:
            type: string
      responses:
        200:
          description: Eliminato il fornitore con l'id specificato
          content: {}
        404:
          description: Fornitore non trovato
          content: {}

# /sede
...
  /fornitoresede/{id}:
    get:
      tags:
      - FornitoreSede
      description: Ottieni un FornitoreSede
      parameters:
        - name: id
          in: path
          description: id del fornitoresede da ottenere
          required: true
          schema:
            type: string
      responses:
        200:
          description: Restituito il FornitoreSede con l'id specificato
          content: {}
        404:
          description: FornitoreSede non trovato
          content: {}
    put:
      tags:
      - FornitoreSede
      description: Aggiorna un ForntoreSede
      parameters:
        - name: id
          in: path
          description: id del fornitoresede da aggiornare
          required: true
          schema:
            type: string
      requestBody:
        description: fornitoresede object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FornitoreSedeBody"
        required: true
      responses:
        200:
          description: Aggiornato il fornitoresede con l'id specificato
          content: {}
        404:
          description: fornitoreSede non trovato
          content: {}
    delete:
      tags:
      - FornitoreSede
      description: Cancella un FornitoreSede
      parameters:
        - name: id
          in: path
          description: l'id del fornitoresede da eliminare
          required: true
          schema:
            type: string
      responses:
        200:
          description: Eliminato il fornitoresede con l'id specificato
          content: {}
        404:
          description: FornitoreSede non trovato
          content: {}
...
  /orario/{id}:
    get:
      tags:
      - Orario
      description: Ottieni un Orario
      parameters:
        - name: id
          in: path
          description: id dell'orario da ottenere
          required: true
          schema:
            type: string
      responses:
        200:
          description: Restituito l'orario con l'id specificato
          content: {}
        404:
          description: Orario non trovato
          content: {}
    put:
      tags:
      - Orario
      description: Aggiorna un Orario
      parameters:
        - name: id
          in: path
          description: id dell' orario da aggiornare
          required: true
          schema:
            type: string
      requestBody:
        description: orario object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrarioBody"
        required: true
      responses:
        200:
          description: Aggiornato l'orario con l'id specificato
          content: {}
        404:
          description: Orario non trovato
          content: {}
    delete:
      tags:
      - Orario
      description: Cancella un Orario
      parameters:
        - name: id
          in: path
          description: l'id dell'orario da eliminare
          required: true
          schema:
            type: string
      responses:
        200:
          description: Eliminato l'orario con l'id specificato
          content: {}
        404:
          description: Orario non trovato
          content: {}
...
  /periodo/{id}:
    get:
      tags:
      - Periodo
      description: Ottieni un Periodo
      parameters:
        - name: id
          in: path
          description: id del periodo da ottenere
          required: true
          schema:
            type: string
      responses:
        200:
          description: Restituito il periodo con l'id specificato
          content: {}
        404:
          description: Periodo non trovato
          content: {}
    put:
      tags:
      - Periodo
      description: Aggiorna un Periodo
      parameters:
        - name: id
          in: path
          description: id del periodo da aggiornare
          required: true
          schema:
            type: string
      requestBody:
        description: periodo object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PeriodoBody"
        required: true
      responses:
        200:
          description: Aggiornato il periodo con l'id specificato
          content: {}
        404:
          description: Periodo non trovato
          content: {}
    delete:
      tags:
      - Periodo
      description: Cancella un Periodo
      parameters:
        - name: id
          in: path
          description: l'id del periodo da eliminare
          required: true
          schema:
            type: string
      responses:
        200:
          description: Eliminato il periodo con l'id specificato
          content: {}
        404:
          description: Periodo non trovato
          content: {}
...
  /presenza/{id}:
    get:
      tags:
      - Presenza
      description: Ottieni una Presenza
      parameters:
        - name: id
          in: path
          description: id della presenza da ottenere
          required: true
          schema:
            type: string
      responses:
        200:
          description: Restituita la presenza con l'id specificato
          content: {}
        404:
          description: Presenza non trovata
          content: {}
    put:
      tags:
      - Presenza
      description: Aggiorna una Presenza
      parameters:
        - name: id
          in: path
          description: id della presenza da aggiornare
          required: true
          schema:
            type: string
      requestBody:
        description: presenza object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresenzaBody"
        required: true
      responses:
        200:
          description: Aggiornata la presenza con l'id specificato
          content: {}
        404:
          description: Presenza non trovata
          content: {}
    delete:
      tags:
      - Presenza
      description: Cancella una Presenza
      parameters:
        - name: id
          in: path
          description: l'id della presenza da eliminare
          required: true
          schema:
            type: string
      responses:
        200:
          description: Eliminata la presenza con l'id specificato
          content: {}
        404:
          description: Presenza non trovata
          content: {}
...
  /servizio/{id}:
    get:
      tags:
      - Servizio
      description: Ottieni un Servizio
      parameters:
        - name: id
          in: path
          description: id del servizio da ottenere
          required: true
          schema:
            type: string
      responses:
        200:
          description: Restituito il servizio con l'id specificato
          content: {}
        404:
          description: Servizio non trovato
          content: {}
    put:
      tags:
      - Servizio
      description: Aggiorna un Servizio
      parameters:
        - name: id
          in: path
          description: id del servizio da aggiornare
          required: true
          schema:
            type: string
      requestBody:
        description: servizio object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServizioBody"
        required: true
      responses:
        200:
          description: Aggiornato il servizio con l'id specificato
          content: {}
        404:
          description: Servizio non trovato
          content: {}
    delete:
      tags:
      - Servizio
      description: Cancella un Servizio
      parameters:
        - name: id
          in: path
          description: l'id del servizio da eliminare
          required: true
          schema:
            type: string
      responses:
        200:
          description: Eliminato il servizio con l'id specificato
          content: {}
        404:
          description: Servizio non trovato
          content: {}
...
components:
  schemas:
...
    FornitoreBody:
      title: Fornitore
      required:
      - nome
      type: object
      properties:    
        nome:
          type: string
          example: mario
        cognome: 
          type: string
          example: rossi
        titolo:
          type: string
          example: dot
        descrizione:
          type: string
          example: descrizione fornitore 
    
  # SedeBody
...
    FornitoreSedeBody:
      title: FornitoreSede
      required:
      - FornitoreId
      - SedeId
      type: object
      properties:
        FornitoreId:
          type: string
          example: 15e6a985-bdb6-441d-b0d8-c42d8f702ea3
        SedeId:
          type: string
          example: 05c8d880-721b-496a-8033-6404e11cf0a5
  
  # FornitoreServizioBody
...
    OrarioBody:
      title: Orario
      required:
      - dalle
      - alle
      - PresenzaId
      - ServizioId
      type: object
      properties:
        dalle:
          type: string
          example: "08:00"
        alle:
          type: string
          example: "12:30"
        descrizione:
          type: string
          example: "descrizione orario"
        PresenzaId:
          type: string
          example: "84605c8f-170e-45f7-b650-4d39a1859981"
        ServizioId:
          type: string
          example: "4b4ba95c-6e58-491c-8e10-c148273cd40e"

  # PeriodoBody
    PeriodoBody:
      title: Periodo
      required:
      - nome
      - dal
      type: object
      properties:
        nome:
          type: string
          example: "Gennaio-Marzo"
        dal:
          type: string
          example: "2022-01-01"
        al:
          type: string
          example: "2022-03-31"
        anticipo:
          type: number
          example: 10
        descrizione:
          type: string
          example: "descrizione periodo"

  # PrenotazioneBody
...
    PresenzaBody:
      title: Presenza
      required:
      - dalle
      - alle
      - giorno
      - FornitoreId
      - PeriodoId
      - SedeId
      type: object
      properties:
        dalle:
          type: string
          example: "14:00"
        alle:
          type: string
          example: "14.30"
        giorno: 
          type: string
          example: "2022-04-14"
        descrizione:
          type: string
          example: "descrizione prenotazione"
        FornitoreId:
          type: string
          example: "15e6a985-bdb6-441d-b0d8-c42d8f702ea3"
        PeriodoId:
          type: string
          example: "05c8d880-721b-496a-8033-6404e11cf0a5"
        SedeId:
          type: string
          example: "9f2cb10f-6880-4bd2-ac6e-5e547bfc83e6"

  # ServizioBody
    ServizioBody:
      title: Servizio
      required:
      - nome
      - durata
      type: object
      properties:   
        nome:
          type: string
          example: visita ortopedica
        durata: 
          type: number
          example: 30
        descrizione:
          type: string
          example: descrizione servizio

  # UtenteBody
...